Don't flicker posframe on self-insert.
This adds a check to the clean-up hook that only hides the posframe if
it was not a `self-insert-command`.

I would like to also skip the hide if commands are any movement or edit
commands. Or potentially if the line is still the same, so those
operations wouldn't also flicker the frame. This is more along the lines
of the original ElDoc displayed in message.